pyproject.toml
[project]
name = "ai-dev"
version = "0.0.1"
description = "AI Agent Development"
authors = [
    { name = "Hulk Pham", email = "tanhunghue@gmail.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.6",
    "langchain-openai>=0.1.22",
    "langchain-anthropic>=0.1.23",
    "langchain>=0.2.14",
    "langchain-fireworks>=0.1.7",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.2.17",
    "tavily-python>=0.4.0",
    "ipython>=8.22.1",
    "graphviz>=0.20.1",
    "sqlalchemy>=2.0.28",
    "alembic>=1.13.1",
    "mysql-connector-python>=8.0.33",
    "langchain-experimental>=0.3.4",
    "google-search-results>=2.4.2"
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"


